+++ packages/banana-react/src/index.ts
@@ -8,5 +8,6 @@ import { Menu } from './menu';
 import { MenuItem } from './menu-item';
 import { Overlay } from './overlay';
 import { Popup } from './popup';
+import { Progress } from './progress';
 
-export { Button, Carousel, Collapse, Countdown, Divider, Dropdown, Menu, MenuItem, Overlay, Popup };
+export { Button, Carousel, Collapse, Countdown, Divider, Dropdown, Menu, MenuItem, Overlay, Popup, Progress };